Refactored large packages
