+++ src-tauri/src/closer.rs
@@ -2,7 +2,7 @@ use serde_json::json;
 use tauri::Window;
 use tracing::info;
 
-pub fn close(window: &Window) {
+pub fn close_win(window: &Window) {
   if let Err(err) = window.hide() {
     info!("Failed to close window: {}", err);
     return;
@@ -11,4 +11,10 @@ pub fn close(window: &Window) {
     info!("Failed to reset state: {}", err);
     return;
   }
+}
+
+#[tauri::command]
+pub fn close(window: tauri::Window) -> Result<(), String> {
+  close_win(&window);
+  Ok(())
 }
