+++ src-tauri/src/page.rs
@@ -66,6 +66,7 @@ impl SettingsDataBuilder {
     map.insert("IMAGE_TO_DATA".into(), "image_data_url".into());
     map.insert("SAVE_BOOKMARKS".into(), "save_bookmarks".into());
     map.insert("GET_CONFIG".into(), "get_config".into());
+    map.insert("VALIDATE_TEMPLATE".into(), "validate_template".into());
     map
   }
 }
